+++ packages/adapters/rig/rugcheck/src/lib.rs
