+++ packages/adapters/rig/rugcheck/src/token_report_detailed.rs
@@ -12,56 +12,48 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::{
-    actions::{fetch_summary_report, TokenCheck},
-    parameters_json_schema,
-};
-use rig::{
-    completion::ToolDefinition,
-    tool::{Tool, ToolEmbedding},
-};
+use solagent_core::{rig::{completion::ToolDefinition, tool::{Tool, ToolEmbedding}}, parameters_json_schema};
 use serde::{Deserialize, Serialize};
-use serde_json::json;
+use solagent_plugin_rugcheck::fetch_detailed_report;
 
 #[derive(Deserialize)]
-pub struct FetchTokenReportSummaryArgs {
+pub struct FetchTokenReportDetailedArgs {
     mint: String,
 }
 
 #[derive(Deserialize, Serialize)]
-pub struct FetchTokenReportSummaryOutput {
-    pub token_check: TokenCheck,
+pub struct FetchTokenReportDetailedOutput {
+    pub token_check: serde_json::Value,
 }
 
 #[derive(Debug, thiserror::Error)]
-#[error("FetchTokenReportSummary error")]
-pub struct FetchTokenReportSummaryError;
+#[error("FetchTokenReportDetailed error")]
+pub struct FetchTokenReportDetailedError;
 
 #[derive(Default)]
-pub struct FetchTokenReportSummary {}
+pub struct FetchTokenReportDetailed {}
 
-impl FetchTokenReportSummary {
+impl FetchTokenReportDetailed {
     pub fn new() -> Self {
-        FetchTokenReportSummary {}
+        FetchTokenReportDetailed {}
     }
 }
 
-impl Tool for FetchTokenReportSummary {
-    const NAME: &'static str = "fetch_summary_report";
+impl Tool for FetchTokenReportDetailed {
+    const NAME: &'static str = "fetch_detailed_report";
 
-    type Error = FetchTokenReportSummaryError;
-    type Args = FetchTokenReportSummaryArgs;
-    type Output = FetchTokenReportSummaryOutput;
+    type Error = FetchTokenReportDetailedError;
+    type Args = FetchTokenReportDetailedArgs;
+    type Output = FetchTokenReportDetailedOutput;
 
     async fn definition(&self, _prompt: String) -> ToolDefinition {
         ToolDefinition {
-            name: "fetch_summary_report".to_string(),
+            name: "fetch_detailed_report".to_string(),
             description: r#"
-                Fetches a summary report for a specific token from RugCheck.
-                
+                Fetches a detailed report for a specific token from RugCheck.
                 Inputs:
-                    - mint: string, the mint address of the token, e.g., "84VUXykQjNvPDm88oT5FRucXeNcrwdQGottJKjkAoqd1" (required).
-                "#.to_string(),
+                - mint: string, the mint address of the token, e.g., "84VUXykQjNvPDm88oT5FRucXeNcrwdQGottJKjkAoqd1" (required).
+            "#.to_string(),
             parameters: parameters_json_schema!(
                 mint: String,
             ),
@@ -69,8 +61,8 @@ impl Tool for FetchTokenReportSummary {
     }
 
     async fn call(&self, args: Self::Args) -> Result<Self::Output, Self::Error> {
-        let token_check = fetch_summary_report(args.mint).await.expect("fetch_summary_report");
-        Ok(FetchTokenReportSummaryOutput { token_check })
+        let token_check = fetch_detailed_report(args.mint).await.expect("fetch_detailed_report");
+        Ok(FetchTokenReportDetailedOutput { token_check })
     }
 }
 
@@ -78,17 +70,17 @@ impl Tool for FetchTokenReportSummary {
 #[error("Init error")]
 pub struct InitError;
 
-impl ToolEmbedding for FetchTokenReportSummary {
+impl ToolEmbedding for FetchTokenReportDetailed {
     type InitError = InitError;
     type Context = ();
     type State = ();
 
     fn init(_state: Self::State, _context: Self::Context) -> Result<Self, Self::InitError> {
-        Ok(FetchTokenReportSummary {})
+        Ok(FetchTokenReportDetailed {})
     }
 
     fn embedding_docs(&self) -> Vec<String> {
-        vec!["Fetches a summary report for a specific token from RugCheck.".into()]
+        vec!["Fetches a summary report for a specific token from FetchTokenReportDetailed.".into()]
     }
 
     fn context(&self) -> Self::Context {}
