+++ packages/plugins/rugcheck/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "solagent-plugin-rugcheck"
+version.workspace = true
+edition.workspace = true
+
+[dependencies]
+serde_json = "1.0"
+serde = { version = "1.0", features = ["derive"] }
+reqwest = { version = "0.12", features = ["json"] }
