+++ packages/plugins/rugcheck/src/lib.rs
@@ -12,7 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub mod solana_token_security_info;
-pub mod token_malicious_info;
-pub mod token_phishing_site_info;
-pub mod token_security_info;
+mod token_report_summary;
+pub use token_report_summary::{fetch_summary_report, TokenCheck};
+
+mod token_report_detailed;
+pub use token_report_detailed::fetch_detailed_report;
+
+pub const RUGCHECK_URL: &str = "https://api.rugcheck.xyz/v1";
