+++ packages/plugins/rugcheck/src/token_report_summary.rs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::primitives::RUGCHECK_URL;
+use crate::RUGCHECK_URL;
 use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Serialize, Deserialize, Default)]
